Finish tooltip tutorial triggers and enable alerts
