+++ src/systems/tutorial.rs
@@ -1,6 +1,10 @@
 use bevy::prelude::*;
 
 use crate::resources::tutorial::TutorialState;
+use crate::ui::UiTag;
+
+#[derive(Component)]
+struct HighlightOriginalColor(Color);
 
 #[derive(Resource)]
 pub struct TutorialUi {
@@ -87,7 +91,7 @@ fn setup_tutorial_ui(mut commands: Commands, asset_server: Res<AssetServer>) {
 }
 
 pub fn tutorial_system(world: &mut World) {
-    let (trigger, title, content, action, step_index, total_steps) = {
+    let (trigger, title, content, action, highlight_tag, step_index, total_steps) = {
         let state = world.resource::<TutorialState>();
         if state.current_step >= state.steps.len() {
             return;
@@ -98,6 +102,7 @@ pub fn tutorial_system(world: &mut World) {
             step.title,
             step.content,
             step.required_action,
+            step.ui_highlight,
             state.current_step,
             state.steps.len(),
         )
@@ -123,6 +128,35 @@ pub fn tutorial_system(world: &mut World) {
         }
         world.entity_mut(container).insert(Visibility::Visible);
 
+        // clear previous highlight
+        if let Some(prev) = world.resource::<TutorialState>().highlighted {
+            if let Some(mut entity) = world.get_entity_mut(prev) {
+                let orig_color = entity
+                    .get::<HighlightOriginalColor>()
+                    .map(|c| c.0);
+                if let (Some(mut bg), Some(orig)) = (entity.get_mut::<BackgroundColor>(), orig_color) {
+                    bg.0 = orig;
+                }
+                entity.remove::<HighlightOriginalColor>();
+            }
+            world.resource_mut::<TutorialState>().highlighted = None;
+        }
+
+        // apply new highlight if defined
+        if let Some(tag) = highlight_tag {
+            let mut q = world.query::<(Entity, &UiTag)>();
+            if let Some((entity, _)) = q.iter(world).find(|(_, t)| t.0 == tag) {
+                if let Some(mut entity_mut) = world.get_entity_mut(entity) {
+                    if let Some(mut bg) = entity_mut.get_mut::<BackgroundColor>() {
+                        let old = bg.0;
+                        bg.0 = Color::YELLOW;
+                        entity_mut.insert(HighlightOriginalColor(old));
+                        world.resource_mut::<TutorialState>().highlighted = Some(entity);
+                    }
+                }
+            }
+        }
+
         if let Some(action_fn) = action {
             action_fn(world);
         }
@@ -134,10 +168,21 @@ fn tutorial_ok_button_system(
     mut interaction_query: Query<&Interaction, (Changed<Interaction>, With<TutorialOkButton>)>,
     tutorial_ui: Res<TutorialUi>,
     mut vis_query: Query<&mut Visibility>,
+    mut state: ResMut<TutorialState>,
+    mut bg_query: Query<(Entity, &mut BackgroundColor, Option<&HighlightOriginalColor>)>,
+    mut commands: Commands,
 ) {
     if let Ok(Interaction::Pressed) = interaction_query.get_single() {
         if let Ok(mut vis) = vis_query.get_mut(tutorial_ui.container) {
             *vis = Visibility::Hidden;
+        }
+        if let Some(entity) = state.highlighted.take() {
+            if let Ok((_, mut color, orig)) = bg_query.get_mut(entity) {
+                if let Some(orig) = orig {
+                    color.0 = orig.0;
+                }
+            }
+            commands.entity(entity).remove::<HighlightOriginalColor>();
         }
     }
 }
